reorganize tiling tools
