+++ packages/datacube-benchmark/src/datacube_benchmark/tiling/__init__.py
@@ -1,0 +1,15 @@
+"""
+tiling
+======
+
+Utilities for working with XYZ map tiles.
+
+Currently provides:
+- get_surrounding_tiles
+- fetch_tile
+from the tiling_utils module.
+"""
+
+from .tiling_utils import get_surrounding_tiles, fetch_tile
+
+__all__ = ["get_surrounding_tiles", "fetch_tile"]
